Encoders and Decoders take property names
INCOMPATIBLE API CHANGE: PReviously, the encoder functions took a
property name and returned an encoder that took an object.  Similarly
for decoders.

This was overly complex; every time the encoder function was used, the
first layer was called.  Remove the currying and simply take the
property name in the encoders and decoders.

* sync_property: change encoder and decoder documentation

* default_encoder, encoders.py, sync_errors, sql_meta_messages: update
  using code