Add DirtyMember to SyncProtocol
Instead of overloading __eq__ and __hash__, add a member DirtyMember
class that represents our information about something we're syncing.
Use a dict instead of a set so we can efficiently get the current
version of a member in the set.

Store an operation alongside DirtyMembers.

Add a test for sync_drain, because that is the operation whose behavior most depends on the areas of code changed.