Use sync_context in receiving objects
Currently the calls to registry.ensure_session are left in place.  The
sync_receive machinery should no longer need these.  The
sync_manager_destinations probably still needs this in some corner
cases.  Also, having some frequent rollbacks to that session will
limit the scope of errors.
Depending on that is sloppy in the long run.