Fix race in testSyncDrained
Adding a done callback to make sure that the second object was not
synchronized provided an opportunity for a race condition.  The object
could be synchronized between the time the future's result was set and
the time that the done callback was called.  Istead mock set_result on
the future wrapping it with a callback.