Message Encoding for Responses
The message format is changed to include a flag int along with the
length.  The lower 16 bits of the flags are critical; sending a not
understood critical flag aborts the connection.  Flag bit 1 is
response needed.

ResponseReceiver tracks what to do with received responses.
DirtyElement gains support for tracking a ResponseReceiver and for
merging ResponseReceivers.

SyncManager.synchronize gains a response parameter and a response_for
parameter.