Working Responses for Delete and Update
* Track outgoing and incoming message numbers in SyncProtocol

* When SyncManager.synchronize is going to send to a destination
  recorded in response_for's ResponseReceiver, we've reached the
  turnover point in flood and have moved from flooding toward the
  decision point to flooding out responses.  At that point we no
  longer want to request future messages get responses.

* In SyncProtocol._send_sync_message, set _MSG_FLAG_RESPONSE_NEEDED
ifwe haven't reached that turnover point and we have a
ResponseReceiver; await a response to that message number.  If we have
a ResponseReceiver and our protocol is recorded in it, then include
_resp_for for all the messages that we are responding to.

* In SyncProtocol._reader_task, if we see _MSG_FLAG_RESPONSE_NEEDED,
create a new ResponseReceiver, tie it to our protocol and this
message, and pass into SyncManager._sync_receive.  Similarly, if the
message has _resp_for, look up the ResponseReceiver and pass into
SyncManager._sync_receive.  It is an error for a message to have the
response needed flag and have _resp_for.  Preliminary stub support for
handling metadata like messages that didn't get a response.

* In ResponseReceiver.merge, don't merge with ourself; duplicates data.

* In all the flood methods in operations.py, include response_for in all calls to synchronize

* in gen_you_have_task, be better about detecting closed protocols and avoid throwing useless exceptions

* New tests for delete and update responses in tests/gateway.py