No Response Handling
When it becomes impossible for a response to be sent, indicate that.
If futures are registered, set their result to None.  If we would
forward a response, then in the next message, include the message
number that will not receive a response.  Trigger sending an empty
message if we need to in order to flush that information.

* SyncProtocol: Support sending an neceiving messages that have no
  _sync_type; any such message should only have meta properties.

* Include _no_resp_for in outgoing messages to indicate other
   previously received messages that will not receive a response.
   Handle and strip on incoming messages.

* ResponseReceiver gets no_response method automatically called on
  __del__ indicating that there will be no response.

* Tests that cover no response flooding, no responses sent alone as
  metadata and no responses combined with other messages.