Errors as Responses
When receiving an object raises a SyncError, propagate that error at
least one hop toward the sender.  If the received message needs a
response, then flood the error as that response.

There is a new class of operations (ResponseOperation) that flood
toward a response tree rather than based on sync_owner.  The 'error'
operation is a concrete instance of this type of operation.

* SyncManager._sync_receive: synchronize out SyncErrors; most errors
  were logged here rather than being raised back to SyncProtocol,
  which already tried to flood SyncErrors.

* All paths that flood SyncErrors: chain the response_for object and use the new error operation.

* SyncError: Significant reworking of how cause, context are handled.
  Track the received javascript generating the error.  Have more
  detail in the string representation.

* incoming forward for SQL objects: reraise
  sqlalchemy.exc.StatementErrors as SyncSqlErrors.

* TestGateway.testErrorResponse: New test to confirm that forwards generating an error flood their error back as a response.