Transition Operation
Provide an operation for fast-changing items such as positions of
graphical elements being moved or resized.

Transitions are intended to inform other SyncManagers of a possible
change as it happens.  When an object exits transition, all parties
are required to restore state as if the transition had not happened.
If the transitioning SyncManager wishes to propose the final state of
the transition as permanent, a sync or forward needs to be
synchronized.

* TransitionTrackerMixin and SqlTransitionTrackerMixin: Classes to add
  transition semantics to SqlSynchronizable and Synchronizable

* operations.transition_operation: new operation.  Transitions are
  never flooded toward the SyncManager initiating them.  Transitions
  preserve response state when forwarding toward the owner if the
  owner is not initiating the transition.  Besides that, transitions
  are flooded to all parties at all states.  It works sort of like the
  transitioning SyncManager has become a temporary owner for the
  object during transition.

* TestGateway.testTransition: new test

* test.utils.settle_loop: It turns out sometimes when data is written
  even on a localhost connection it is not immediately available.
  Wait 1/100 of a second in select for the event list to settle when
  determining if a loop has settled.

* Rename SyncSqlError to SqlSyncError for consistency with other types