Fix owner transitions
When an object owner transitions an object, it is unlikely to get a
response especially in the situation where there are no receivers for
the object.  This tends to pull the object out of transition, which
can be problematic.  If the future returns a None result, don't pull
local objects out of transition.

While making this change, I noticed the call to (the previous
transition's) remove_from_transition in store_for_transition was
broken; fix it.