cert_hash becomes dest_hash
A cert_hash is only one kind of dest_hash.  In most of the code prefer
dest_hash (and DestHash and SqlDestHash) to cert_hash.

This permits non-certificate destinations like web sockets or unix
sockets.

We prefer cert_hash when we know we're dealing with certificates.  So bin/entanglement-cli prefers cert_hash as do SyncManager.cert_hash from the local certificate.