Correctly partition for remove_from_transition
In the previous implementation of transitions_partitioned, then
partitioning was only applied while _sync_receive was running.
However, when a response is received, callbacks can be scheduled based
on the futures.  These call remove_from_transition, which doesn't have
the right dictionary as transition_tracked_objects, so the object is
not removed from transition.  As a result, tests erroneously passed.

Now, override each instance's idea of transition_tracked_objects when
an instance is stored for transition inside _sync_receive so that
remove_from_transition will always work on that instance.