Make SqlSyncDestination polymorphic
Allow SqlSyncDestination to be polymorphic

Don't include owners we have no plan to send in the MyOwners message.

The cls parameter of sync_manager_destinations has changed
functionality.  Previously it coerced destinations to be the given
class.  Now it restricts the query.  The test expects the old
behavior; I'm not entire sure which behavior is right.  For now I've
marked the test as an expected failure.