No autoflush while removing objects from session
SqlTransitionTrackerMixin needs to disable autoflush while removing
objects from session.  If a sync_owner or destination wer expired, the
flush could flush out modifications that were intended to be sent in a
transition.  First, this could be bad because those modification might
be more permanent than intended.  Also, it affected the set of
apparently modified attributes.