SyncOwner.destination
It's no longer the case that a SyncOwner's destination is guaranteed
to be an SqlSyncDestination.  So, rename the destination relationship
to sql_destination.  This relationship may be used when an application
knows that a SqlSyncDestination is used.

On an incoming sync we set destination to be the actual sender of the
masseg.  This destination object is not likely to be in the same
session as the SyncOwner and may or may not be an SqlSyncDestination.
This is intended for backward compatibility with existing
applications.

Remove destination logic from the after_flush handler that is no longer appropriate; same for SqlTransitionTrackerMixin.

Update tests.