DestHash inequality
DestHash didn't have an __ne__ method so it was possible for a ==b and
a != b to both be true comparing a desthash to its string
representation.

Add a test for desthash equality.

In splitting SyncDestination out into SyncDestinationBase, the string
form of a DestHash is not converted to a DestHash in the constructor;
fix.