Encoders for Float and Integer
Without encoders for Float and Integer, if an object's numeric columns
are set to string representations of the numbers, then this is
synchronized over the network.  Sqlalchemy gets numeric contexts
because it refreshes from the database.  Entanglement uses encoders
for similar concepts.

Add encoders for Integer and Float

Add a test to confirm that a float set as a string is sent as a float.