SyncRegistry: adjust unregistered classes on register
We typically want applications rather than libraries to provide schemas.  In the library then, the schema is not available to extend.
So provide a way that library classes can inherit the schema:

If a class is registered with a registry that extends Synchronizable
but does not have syncType, see if its name is in the bases.  If so,
then mix in syncType, _syncAttributes and syncPrimaryKeys and register
the class.