relationships: missing_node
Sometimes we don't have referential integrity and we'd like a pseudo
object to fill in for an object that hasn't been received, perhaps not
created, or perhaps we aren't in the should_send.

Add support to that to the relationship logic.

Also, fix a bug in the remote_prop getter.

Add tests for the above