Return empty nodes on no key
Previously we returned undefined in the "many" side of a one-to-many relationship had no entries.
In general we'd rather return a list.

We don't have great choices.  We can allocate a list the first time we
see a key, which is a resource leak.  Alternatively we can return a
fresh empty list until we get members of the relationship.  Returning
a fresh empty list can be bad for reactivity systems like Vue.

For now, allocate the node whenever we see the first get, favoring Vue against resource leaks.