Try to use constructor when we can
Originally syncConstruct bypassed using the constructor.  They The
theory was that a constructor might need arguments and might do the
wrong thing for objects being restored from serialization.  Classes
that wanted the behavior could override.

In practice this proved hard for descendants of
PersistentSynchronizable.  Those classes might want to override the
behavior for new objects but not for looking up objects in the map.
In addition, the behavior proved highly problematic for UI -focused
applications like most browser apps.  The UI is going to want to
maintain state that is constructed from what is synchronized.  This is
true for Photon, for the desktop sender and for my test app.  In all
these cases, attributes and state needs to be set up on the object.
For Vue2 apps the reactivity implications of not setting up these
attributes are particularly insidious.

So, to improve:

* Split out empty object construction into syncEmptyObject static method.

* By default try the constructor; if it throws produce an empty
object.  This has both false positive and false negatives.  It's a
best guess.  If it is wrong for a given class, override
syncEmptyObject.