Changes for Sqlalchemy 1.4
Most of these are changes in where symbols live, compatible back with
1.3 as well.

Don't cascade session merges for SyncOwner. We ran into situations in
the tests where old values for serial numbers were trying to get
flushed for nonlocal objects.  This could have been fixed in the
tests, but not merging SyncOwner appears to be the principle of least
surprise here, and appears not to break any of our existing usages.